move context keys to sdk audit
