+++ sdk/audit/context_keys.go
@@ -1,0 +1,9 @@
+package audit
+
+type AuditContextKey string
+
+const (
+	RequestIDContextKey AuditContextKey = "request-id"
+	UserAgentContextKey AuditContextKey = "user-agent"
+	ActorIDContextKey   AuditContextKey = "actor-id"
+)
